profile-settings-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
httpx>=0.28.0
Pillow>=10.4.0
aioboto3>=13.3.0
eval_type_backport>=0.2.0
